ktfirsttry/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
Django==3.0.6
django-bootstrap==0.2.4
django-bootstrap4==1.1.1
django-celery==3.3.1
django-compat==1.0.15
django-jquery==3.1.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
kombu==4.6.8
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
rabbitmq==0.2.0
redis==3.5.2
soupsieve==2.0.1
sqlparse==0.3.1
